refactor: move hook
