+++ src/pages/scheduler/components/SchedulerCalendar.tsx
@@ -10,6 +10,7 @@ import { useDarkMode } from 'lib/hooks/useDarkMode';
 
 import { CalendarEvent } from 'pages/scheduler/components/Event';
 import { CalendarToolBar } from 'pages/scheduler/components/Toolbar';
+import { useInitialDateTime } from 'pages/scheduler/hooks/useInitialDatetime';
 import { coursesToVCalendar } from 'pages/scheduler/shared/exporter';
 import { CreateEventsFromCourses } from 'pages/scheduler/shared/parsers';
 import { CustomEvent } from 'pages/scheduler/shared/types';
@@ -25,29 +26,6 @@ const localizer = dateFnsLocalizer({
   getDay,
   locales: { 'en-US': enUS },
 });
-
-const useInitialDateTime = (term?: string) => {
-  const today = useMemo(() => new Date(), []);
-  // determine what date to position the calendar on.
-  return useMemo<Date>(() => {
-    // if there is a term, use today as the default
-    if (term === undefined) return today;
-
-    // eg. 202105 => 2021, 05
-    const month = parseInt(term.substring(4, 6));
-    const year = parseInt(term.substring(0, 4));
-    // if the selected term's first month is before or during the current month
-    // && the selected term's last month is after or during the current month
-    // && the selected term's year is the current year
-    // return today
-    // else return the 2nd week (when most labs start) of the first month of the selected term
-    if (month <= today.getMonth() + 1 && month + 4 >= today.getMonth() + 1 && year === today.getFullYear()) {
-      return today;
-    } else {
-      return new Date(year, month - 1, 12);
-    }
-  }, [today, term]);
-};
 
 export interface SchedulerCalendarProps {
   /**
